guard against bad zip codes. remove function creation from app.py and put into own .py file
